seperate tests into ce and ent
