+++ command/command_testonly/operator_usage_testonly_oss_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: BUSL-1.1
 
-//go:build testonly
+//go:build testonly && !enterprise
 
 package command_testonly
